feat: custom color picker  and some stuff
